requirements.txt
ultralytics
onnxruntime-gpu
pillow==10.2.0
pandas
numpy
jsonpath
lxml
PyExecJS2
requests
loguru
killport
email-validator
fastapi==0.110.0
fastapi-restful==0.5.0
pydantic==2.6.3
uvicorn==0.27.1